+++ src/alerts/alert.component.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Alert as BSAlert } from 'react-bootstrap';
+import { Alert } from 'react-bootstrap';
 import { Icon } from '@opuscapita/react-icons';
 import { OCAlert as OCAlertActions } from './alerts.actions';
 
@@ -38,11 +38,11 @@ export class OCAlert extends React.PureComponent {
 
   render() {
     return (
-      <BSAlert bsStyle={this.props.type} onDismiss={this.handleAlertDismiss}>
+      <Alert bsStyle={this.props.type} onDismiss={this.handleAlertDismiss}>
         <div className="alert-content">
           { this.getIcon() }
           <span>{ this.props.message }</span>
         </div>
-      </BSAlert>);
+      </Alert>);
   }
 }